+++ homepage/example.js
@@ -124,9 +124,9 @@ function makePanel(e) {
 function measureDisplaySize(e) {
   RemoteCalibrator.displaySize(displayData => {
     printMessage(
-      `Display size is ${displayData.value.displayWidthPx}px in width and ${
+      `Display size is ${displayData.value.displayWidthPx} px in width and ${
         displayData.value.displayHeightPx
-      }px in height, measured at ${parseTimestamp(displayData.timestamp)}.`
+      } px in height, measured at ${parseTimestamp(displayData.timestamp)}.`
     )
 
     changeClass(e.target, 'complete')
@@ -141,9 +141,9 @@ function measureDisplaySize(e) {
 function measureScreenSize(e) {
   RemoteCalibrator.screenSize({}, screenData => {
     printMessage(
-      `Screen size is ${screenData.value.screenDiagonalIn}in [Width: ${
+      `Screen size is ${screenData.value.screenDiagonalIn} in [Width: ${
         screenData.value.screenWidthCm
-      }cm, Height: ${screenData.value.screenHeightCm}cm, PPI: ${
+      } cm, Height: ${screenData.value.screenHeightCm} cm, PPI: ${
         screenData.value.screenPpi
       }, PPI (Physical): ${
         screenData.value.screenPhysicalPpi
@@ -158,7 +158,7 @@ const measureDistanceCallback = distanceData => {
   printMessage(
     `The viewing distance is ${
       distanceData.value
-    }cm, measured at ${parseTimestamp(distanceData.timestamp)}, by ${
+    } cm, measured at ${parseTimestamp(distanceData.timestamp)}, by ${
       distanceData.method
     } method.`
   )
@@ -199,11 +199,11 @@ function trackViewingDistance(e) {
       trackP.innerHTML = gotData(
         `The dynamic viewing distance is ${
           data.value.viewingDistanceCm
-        }cm at ${parseTimestamp(data.timestamp)}, measured by ${
+        } cm at ${parseTimestamp(data.timestamp)}, measured by ${
           data.method
-        } method. The near point is at [${data.value.nearPointCm.x}cm, ${
+        } method. The near point is at [${data.value.nearPointCm.x} cm, ${
           data.value.nearPointCm.y
-        }cm] compared to the center of the screen.`
+        } cm] compared to the center of the screen.`
       )
     }
   )
@@ -298,9 +298,9 @@ function trackGaze(e) {
     },
     data => {
       gazeP.innerHTML = gotData(
-        `The gaze position is [${data.value.x}px, ${
+        `The gaze position is [${data.value.x} px, ${
           data.value.y
-        }px] at ${parseTimestamp(data.timestamp)}.`
+        } px] at ${parseTimestamp(data.timestamp)}.`
       )
     }
   )