+++ src/const.js
@@ -1,0 +1,8 @@
+import RemoteCalibrator from './core'
+
+RemoteCalibrator.prototype._CONST = Object.freeze({
+  CREDIT_TEXT: {
+    BLIND_SPOT_TEST: `As suggested by the Li et al. (2020) "Virtual Chinrest" paper.`,
+    CREDIT_CARD: `Credit card suggested by the Li et al. (2020) "Virtual Chinrest" paper.`,
+  },
+})