+++ src/distance/distance.js
@@ -15,6 +15,7 @@ import {
 } from '../components/onCanvas'
 import { bindKeys, unbindKeys } from '../components/keyBinder'
 import { addButtons } from '../components/buttons'
+import { soundFeedback } from '../components/sound'
 import text from '../text.json'
 
 const blindSpotHTML = `<canvas id="blind-spot-canvas"></canvas>`
@@ -39,8 +40,9 @@ export function blindSpotTest(RC, options, toTrackDistance = false, callback) {
   RC.background.appendChild(blindSpotDiv)
   RC._constructFloatInstructionElement(
     'blind-spot-instruction',
-    `Keep your <span id="eye-side"></span> eye closed and focus on the cross.`
+    `Close <span id="eye-side"></span> eye.`
   )
+  RC._addCreditOnBackground(RC._CONST.CREDIT_TEXT.BLIND_SPOT_TEST)
 
   // Get HTML elements
   const c = document.querySelector('#blind-spot-canvas')
@@ -48,6 +50,7 @@ export function blindSpotTest(RC, options, toTrackDistance = false, callback) {
 
   const eyeSideEle = document.getElementById('eye-side')
   let eyeSide = (eyeSideEle.innerText = 'LEFT').toLocaleLowerCase()
+  RC._setFloatInstructionElementPos(eyeSide, 16)
   let crossX = _getCrossX(eyeSide, c.width)
 
   let circleBounds
@@ -81,6 +84,8 @@ export function blindSpotTest(RC, options, toTrackDistance = false, callback) {
 
   // SPACE
   const finishFunction = () => {
+    soundFeedback()
+
     tested += 1
     // Average
     dist.push(
@@ -113,6 +118,8 @@ export function blindSpotTest(RC, options, toTrackDistance = false, callback) {
       if (eyeSide === 'left')
         eyeSide = (eyeSideEle.innerText = 'RIGHT').toLocaleLowerCase()
       else eyeSide = (eyeSideEle.innerText = 'LEFT').toLocaleLowerCase()
+      RC._setFloatInstructionElementPos(eyeSide, 16)
+
       circleBounds = _getCircleBounds(eyeSide, crossX, c.width)
       circleX = circleBounds[eyeSide === 'left' ? 0 : 1]
       v = eyeSide === 'left' ? 1 : -1