+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `nextHeadline` and `nextDescription` in panel options to change panel text after finishing all calibrations.
+
+### Fixed
+
+- Dependency security issues.
+
 ## [0.2.1] - 2021-09-14
 
 Minor fixes and updates.