+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2]
+
 ### Added
 
 - Sound feedback (2000 Hz, 0.05 s) for responses in blind spot test.
@@ -200,7 +202,8 @@ No new feature updates in this release. Updated dependency packages and the lice
 
 The framework and some basic functions, e.g., screen size calibration. Released for integration testing.
 
-[unreleased]: https://github.com/EasyEyes/remote-calibrator/compare/v0.2.1...develop
+[unreleased]: https://github.com/EasyEyes/remote-calibrator/compare/v0.2.2...develop
+[0.2.2]: https://github.com/EasyEyes/remote-calibrator/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/EasyEyes/remote-calibrator/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/EasyEyes/remote-calibrator/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/EasyEyes/remote-calibrator/compare/v0.1.0...v0.1.1